+++ deployer/default/deploy/task/deploy_fast.php
@@ -15,6 +15,9 @@ task('deploy-fast', [
     // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-composer-install
     'deploy:check_composer_install',
 
+    // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-composer-validate
+    'deploy:check_composer_validate',
+
     // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-branch-local
     'deploy:check_branch_local',